requirements.txt
aiogram==3.0.0b7
asyncpg==0.27.0
msgpack~=1.0.5
nats-py==2.1.7